+++ dao/src/main/java/entities/MedWorker.java
@@ -11,7 +11,7 @@ public abstract class MedWorker extends Person {
     }
 
     public MedWorker(String firstName, String secondName, int age, Sex sex,
-                     Educations education, int experience, int userId) {
+                     Educations education, int experience, Long userId) {
         super(firstName, secondName, age, sex, userId);
         this.education = education;
         this.experience = experience;
